+++ packages/pg-dev/src/typed-pg-vitest-setup/index.ts
@@ -1,5 +1,0 @@
-import { afterAll, beforeEach } from 'vitest'
-
-import { setupTypedPgVitest } from '../setup-helper'
-
-setupTypedPgVitest({ afterAll, beforeEach, projectRoot: process.cwd() })